feat: #36 member service에 기본 이미지는 삭제하지 않는 조건 추가

+++ src/main/java/com/example/sabujak/member/service/MemberService.java
@@ -13,6 +13,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import static com.example.sabujak.security.exception.AuthErrorCode.ACCOUNT_NOT_EXISTS;
@@ -27,6 +28,8 @@ public class MemberService {
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
     private final ImageService imageService;
     private final MemberImageRepository memberImageRepository;
+
+    private final String defaultMemberImageUrl = "https://bzbz-file-bucket.s3.ap-northeast-2.amazonaws.com/Member-Default-Image.png";
 
     public MemberResponseDto.AllInformation getMyInfo(String email) {
         final Member member = memberRepository.findWithCompanyAndImageByMemberEmail(email)
@@ -81,7 +84,7 @@ public class MemberService {
         member.setImage(memberImage);
         memberImage.setMember(member);
 
-        if (exImage != null && exImage.getImageUrl() != null)  //이전 사용자 사진 삭제
+        if (exImage != null && exImage.getImageUrl() != null && !exImage.getImageUrl().equals(defaultMemberImageUrl))  //이전 사용자 사진 삭제
             imageService.deleteImage(exImage.getImageUrl());
 
     }
